added get_envp
