+++ src/builtins/export/export.c
@@ -6,7 +6,7 @@
 /*   By: spenning <spenning@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 14:59:21 by spenning          #+#    #+#             */
-/*   Updated: 2024/06/27 15:12:39 by spenning         ###   ########.fr       */
+/*   Updated: 2024/06/28 14:52:03 by spenning         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,55 +21,6 @@
 // export lol=wat wat wat
 // export wat=lol=wat {wat='lol=wat'}
 // export wat= {wat="", echo $wat = ""}
-
-char	*export_get_key(char *input)
-{
-	int		index;
-	char	*key;
-
-	index = 0;
-	while (input[index] != '\0')
-	{
-		if (input[index] == '=')
-			break ;
-		index++;
-	}
-	index++;
-	key = ft_calloc(index, 1);
-	if (key == NULL)
-		return (NULL);
-	ft_strlcpy(key, input, index);
-	return (key);
-}
-
-char	*export_get_value(char *input)
-{
-	int		start;
-	int		end;
-	int		len;
-	char	*value;
-
-	start = 0;
-	end = 0;
-	len = ft_strlen(input);
-	while (input[start] != '=' && input[start] != '\0')
-		start++;
-	if (start == len)
-	{
-		value = ft_strdup("");
-		if (value == NULL)
-			return (NULL);
-		return (value);
-	}
-	end = ++start;
-	while (input[end] != '\0')
-		end++;
-	value = ft_calloc(((end - start) + 1), 1);
-	if (value == NULL)
-		return (NULL);
-	ft_strlcpy(value, input + start, (end - start) + 1);
-	return (value);
-}
 
 char	*export_check_value(char *value)
 {
@@ -103,12 +54,12 @@ void	export(t_msdata *data)
 
 	if (double_array_len(data->argv) == 1)
 		return (env(data));
-	key = export_get_key(data->argv[1]);
+	key = get_envp_key(data->argv[1]);
 	if (key == NULL)
-		ms_error("malloc error in export_get_key");
-	value = export_get_value(data->argv[1]);
+		ms_error("malloc error in envp_get_key in export");
+	value = get_envp_value(data->argv[1]);
 	if (value == NULL)
-		ms_error("malloc error in export_get_value");
+		ms_error("malloc error in envp_get_value in export");
 	value = export_check_value(value);
 	if (value == NULL)
 		ms_error("malloc error in export_check_value");
